Centralize XML attribute normalization
Move shared attribute extraction into a crate-level helper used by both
spoken-text and streaming parser paths. Document the XML 1.0 assumption at
that helper so future XML version handling has a single point to change.
